moved user avatar to it's own component along with its state and functions
